Add GetAllBlogs + GetAllPages script
